[hotfix] Change event API back to float due to failing tests

+++ streampipes-model/src/main/java/org/apache/streampipes/model/runtime/Event.java
@@ -189,7 +189,7 @@ public class Event {
     }
   }
 
-  public void addField(String runtimeName, double value) {
+  public void addField(String runtimeName, Float value) {
     addPrimitive(runtimeName, value);
   }
